requirements.txt
alembic==1.6.4
bcrypt==3.2.0
cffi==1.14.5
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Migrate==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
psycopg2==2.8.6
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.15
Werkzeug==2.0.1
WTForms==2.3.3
